LAB-COMPETITIVE-ANALYSIS/data/LAB-COMPETITIVE-ANALYSIS/pyproject.toml
[tool.poetry]
name = "lab-competitive-analysis"
version = "0.1.0"
description = ""
authors = ["khalilchikhaoui <ge78fin@mytum.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
kedro = {extras = ["all"], version = "^0.19.11"}
ipython = "^8.30.0"
kedro-viz = "^10.1.0"
kedro-datasets = "^5.1.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
scikit-learn = "^1.6.0"
openpyxl = "^3.1.5"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pyarrow = "^18.1.0"
fastparquet = "^2024.11.0"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
trulens = "^1.3.3"
trulens-providers-openai = "^1.3.3"
trulens-providers-huggingface = "^1.3.3"
bert-score = "^0.3.13"
evaluate = "^0.4.3"
absl-py = "^2.1.0"
rouge-score = "^0.1.2"
pandas = "^2.2.3"
tenacity = "^9.0.0"
langchain-community = "^0.3.16"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.kedro]
package_name = "lab_competitive_analysis" #subdirectory_name
project_name = "LAB-COMPETITIVE-ANALYSIS" #rootdirectory_name
kedro_init_version = "0.19.10" #same kedro version in depandency
tools = [ "Data Structure", "Kedro Viz",]
example_pipeline = "True"
source_dir = "lab_competitive_analysis" #subdirectory_name

[tool.kedro_telemetry]
project_id = "26d5f9c187384ee5884f23d1105ed437"
